Add files via upload
Add "consts.mat" which is consist of the constants which are repeatedly used in the codes. Users can easily modify "consts.mat" to switch to other values rather than find and replace these constants in the scripts.
Add "custom_mat_files.m" for generation of custom "e.mat" and "expo_age.mat".